Part #5 + correction
Includes a correction to workaround that seems not had been nocessary.
I.e. VotingServicecodePkg.Code used to imply 1.0.2 and not 1.0.1 which
must had been an error.

This patch seems to imply getting the string from FabricRuntime in
runtime, which is surely much better but which not necessarily part of
the work of making the soulution stateful (examination TBD).